requirements.txt
amqp==2.1.4
appdirs==1.4.2
billiard==3.5.0.2
celery==4.0.2
Django==1.10.5
django-celery==3.2.1
django-filter==1.0.1
django-mailing==0.1.3
djangorestframework==3.5.4
drf-nested-routers==0.11.1
kombu==4.0.2
Markdown==2.6.8
packaging==16.8
pyparsing==2.1.10
pytz==2016.10
redis==2.10.5
six==1.10.0
vine==1.1.3
